requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pydantic==2.5.0
python-dotenv==1.0.0
alembic==1.13.1
python-multipart==0.0.6
jinja2==3.1.2
aiofiles==23.2.1
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
redis==5.0.1
celery==5.3.4